requirements.txt
opencv-python
tqdm
scipy
pyyaml